pyproject.toml
[project]
name = "robotframework-benchmark"
version = "0.1.0"
description = "A benchmark library for performance testing Robot Framework core implementations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.13"
dependencies = [
    "robotframework>=7.0",
    "psutil>=5.9",
    "memory-profiler>=0.61",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-benchmark>=5.1",
    "ruff>=0.5",
]

[project.scripts]
rfbenchmark = "robotframework_benchmark.cli:main"

[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
